Move diff to dev
